feat(settle): 将价格字段从BigDecimal改为Int类型
涉及商品、优惠券及结算相关模块的价格字段类型统一变更，包括SkuEntity、CouponEntity及各类结算BO和View对象，简化金额计算逻辑。

+++ src/main/kotlin/com/hello/hello_store_service/model/entity/activity/CouponEntity.kt
@@ -12,8 +12,8 @@ data class CouponEntity(
     val description: String,
     val type: CouponType,
     val tag: String?,
-    val threshold: BigDecimal,     // 使用门槛
-    val discount: BigDecimal?,      // 减免金额
+    val threshold: Int,     // 使用门槛
+    val discount: Int?,      // 减免金额
     val discountRate: Float?,  // 折扣券
     val validDays: Int = 0      // 有效天数
 )
